pyproject.toml
[project]
name = "invoice_app"
version = "0.1.0"
description = "Invoice Application"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "pydantic==2.4.2",
    "pydantic-settings==2.0.3",
    "reportlab==4.0.4",
    "python-multipart==0.0.6",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "bcrypt==4.0.1",
    "beanie==1.25.0",
    "motor==3.3.2",
    "python-dotenv==1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."] 